vringh: host-side implementation of virtio rings.
	Changes since v1:
	1) Fix for broken race handling with event indexes.
	2) Fix backward vringh_notify_enable_user / vringh_notify_enable_kern.
	3) iovec/kvec iterators mangle iov/kvec so it's always usable.

Getting use of virtio rings correct is tricky, and a recent patch saw
an implementation of in-kernel rings (as separate from userspace).

This abstracts the business of dealing with the virtio ring layout
from the access (userspace or direct); to do this, we use function
pointers, which gcc inlines correctly.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>
Acked-by: Michael S. Tsirkin <mst@redhat.com>